Remove middleman method and allow previews for internal URLs
