+++ packages/editor/src/components/provider/use-block-editor-settings.js
@@ -6,27 +6,19 @@ import { pick, defaultTo } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { Platform, useMemo, useCallback } from '@wordpress/element';
+import { Platform, useMemo } from '@wordpress/element';
 import { useDispatch, useSelect } from '@wordpress/data';
 import {
 	store as coreStore,
 	__experimentalFetchLinkSuggestions as fetchLinkSuggestions,
 	__experimentalFetchUrlData as fetchUrlData,
 } from '@wordpress/core-data';
-import {
-	getAuthority,
-	isURL,
-	getProtocol,
-	isValidProtocol,
-} from '@wordpress/url';
 
 /**
  * Internal dependencies
  */
 import { mediaUpload } from '../../utils';
 import { store as editorStore } from '../../store';
-
-const EMPTY_DATA = {};
 
 /**
  * React hook used to compute the block editor settings to use for the post editor.
@@ -41,8 +33,6 @@ function useBlockEditorSettings( settings, hasTemplate ) {
 		reusableBlocks,
 		hasUploadPermissions,
 		canUseUnfilteredHTML,
-		baseUrl,
-		hasResolvedLocalSiteData,
 	} = useSelect( ( select ) => {
 		const { canUserUseUnfilteredHTML } = select( editorStore );
 		const isWeb = Platform.OS === 'web';
@@ -75,54 +65,6 @@ function useBlockEditorSettings( settings, hasTemplate ) {
 	}, [] );
 
 	const { undo } = useDispatch( editorStore );
-
-	// Temporary home - should this live in `core-data`?
-	const fetchRichUrlData = useCallback(
-		function ( url, fetchOptions = {} ) {
-			if ( ! isURL( url ) ) {
-				return Promise.reject(
-					new TypeError( `${ url } is not a valid URL.` )
-				);
-			}
-
-			// Test for "http" based URL as it is possible for valid
-			// yet unusable URLs such as `tel:123456` to be passed.
-			const protocol = getProtocol( url );
-
-			if (
-				! isValidProtocol( protocol ) ||
-				! protocol.startsWith( 'http' ) ||
-				! /^https?:\/\/[^\/\s]/i.test( url )
-			) {
-				return Promise.reject(
-					new TypeError( `${ url } does not have a valid protocol.` )
-				);
-			}
-
-			// If the baseUrl is still resolving then return
-			// empty data for this request.
-			if ( ! hasResolvedLocalSiteData ) {
-				return Promise.resolve( EMPTY_DATA );
-			}
-
-			// More accurate test for internal URLs to avoid edge cases
-			// such as baseURL being included as part of a query string
-			// on the target url.
-			const baseUrlAuthority = getAuthority( baseUrl );
-			const urlAuthority = getAuthority( url );
-
-			const isInternal = urlAuthority === baseUrlAuthority;
-
-			// Don't handle internal URLs (yet...).
-			if ( isInternal ) {
-				return Promise.resolve( EMPTY_DATA );
-			}
-
-			// If external then attempt fetch of data.
-			return fetchUrlData( url, fetchOptions );
-		},
-		[ baseUrl, hasResolvedLocalSiteData ]
-	);
 
 	return useMemo(
 		() => ( {
@@ -170,7 +112,7 @@ function useBlockEditorSettings( settings, hasTemplate ) {
 			__experimentalReusableBlocks: reusableBlocks,
 			__experimentalFetchLinkSuggestions: ( search, searchOptions ) =>
 				fetchLinkSuggestions( search, searchOptions, settings ),
-			__experimentalFetchRichUrlData: fetchRichUrlData,
+			__experimentalFetchRichUrlData: fetchUrlData,
 			__experimentalCanUserUseUnfilteredHTML: canUseUnfilteredHTML,
 			__experimentalUndo: undo,
 			outlineMode: hasTemplate,
